requirements.txt
<<<<<<< HEAD
Flask
torch
torchvision
scikit-learn
numpy
Pillow
requests
gunicorn
=======
Flask
torch
torchvision
scikit-learn
numpy
Pillow
requests
gunicorn
>>>>>>> blackboxai/upload-project
